requirements.txt
Flask==3.1.1
flask-cors==6.0.1
pycryptodome==3.23.0